requirements.txt
Flask
flask-cors
tensorflow
numpy
pillow
scikit-image
gdown
